Fix a crash, probably caused by a change in Qt.
Steps to reproduce the crash (tested on Kubuntu 10.04, with updates as of today):
Enter "forward 100" in the code editor, press run. Replace the "forward 100" with "clear", press run, crash.

Qt seems to delete the wrong item from the scene. Setting the index method to the default fixes the problem and doesn't seem to result in any
noticable performance drawbacks. If time permits maybe have a look at this in the 4.5 timeframe.

svn path=/trunk/KDE/kdeedu/kturtle/; revision=1070113