Speed up executing by disabling the inspector when running in full speed mode without highlighting
When a user doesn't want to see the highlights in the editor, there is a good change he or she
also doesn't care for the information in the inspector. This commit changes the
'full speed no highlighting'-speed to also disable the inspector. This can be a huge speedup
for some scripts.

svn path=/trunk/KDE/kdeedu/kturtle/; revision=1107083